Implement temporal logic to express liveness propertiess
This commit is an attemp to solve #3 but it is still work in progress.

Since Verus does not support temporal logic which is necessary to express and prove
liveness properties, we need to implement temporal logic support with Verus to be
able to reason about an execution, instead of a single state.

This consists of mainly two tasks: (1) implement `lift-state` to transform state predicates
to temporal predicates that work on an infinite sequence of states and (2) implement operators
like always [], eventually <>, leads to ~> and so on.

Currently, we implement `lift_state` as a spec function which takes a closure (state predicate)
and returns a closure (temporal predicate). Similarly, the temporal logic operators are functions
that take one or two closures (temporal predicates) and return one closure (temporal predicate).

Unfortunately, this implementation does not compile because Verus does not support returning closure
type very well. We need to discuss with Verus team and figure out the best way to implement temporal logic.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>