Implement compound state machine (#48)
Implement the compound state machine that consists of a Kubernetes API state machine, a controller state machine, a client state machine, and a network state machine. Generalize the enabled proof and separate out the liveness proof that is not specific to the controller.

The biggest difference between this compound state machine and the summer school examples is that: The summer school examples choose the entire `MessageOp` which includes `recv`(`Option<Message>`) and `send`(`Set<Message>`), while the compound state machine only chooses the `recv` -- message to deliver in each step in `next_step`. To make the host and network agree on what messages are sent, in each action the host returns its output messages and the network checks these messages
```
pub open spec fn kubernetes_api_action(recv: Option<Message>) -> ActionPred<CompoundState> {
    |s: CompoundState, s_prime: CompoundState| {
        &&& kubernetes_api::next(recv, s.kubernetes_api_state, s_prime.kubernetes_api_state)
        &&& network::next(recv, s.network_state, s_prime.network_state, kubernetes_api::output(recv, s.kubernetes_api_state, s_prime.kubernetes_api_state))
        ...
    }
}
```

This is also why we need a `output` in each `HostAction`.

I made this decision mainly to make it easier to write `*_enabled` proofs. If the compound action is parameterized by `MessageOp`, then in the liveness proof when invoking `*_enabled` we also need to feed the entire `MessageOp` including the messages the host sends to the network. I feel it is counter-intuitive because reasoning about enabled should only require us to think about the precondition of the action, rather than the output/transition of the action.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>