Prove liveness of simple controller (#54)
### Overview

This new example is developed based on the compound state machine example. It includes the spec of a simple controller and the Kubernetes environment, and it proves the liveness property of the simple controller's reconcile logic. All the specs are under examples/kubernetes_cluster/spec and all the proofs are under examples/kubernetes_cluster/proof.

### What does the spec do

The top-level distributed system spec consists of a controller, Kubernetes API, client, and network spec. The client and network specs are the same as the compound state machine example.

The Kubernetes API spec has only one action: handle_request. This action receives a request and does a lot of things, including updating object store (at etcd), returning a response of the request, triggering built-in controllers with the new updates in the cluster state.

The controller spec describes a notification-based level-triggering controller. The controller maintains two key data structures -- a map, ongoing_reconciles, that stores the local reconcile state per CR key, including the pending API request, and a set, scheduled_reconciles, that contains all the CR keys whose reconciles are requeued. The controller has four actions: (1) trigger_reconcile, which receives a relevant notification (also called event message in the spec) that triggers a reconcile for a particular custom resource CR and initializes the reconcile local state for this CR's key, (2) run_rescheduled_reconcile, which picks one of the CR key in the scheduled_reconciles and initializes the reconcile local state for this CR's key, (3) continue_reconcile, which picks a CR key that is in reconcile, receives the response for the pending request of the previous step (if any), runs the next step in the user-provided reconcile_core function, updates the local reconcile state for this CR key and also the new pending request (if any), (4) end_reconcile, which is only enabled when the reconcile_core is done. It removes the local state for this CR key and reschedules a new reconcile for this CR key.

Note that the user should provide a reconcile definition to complete the controller spec, including (1) reconcile_core, the controller-specific reconcile function. (2) reconcile_trigger, a function that decides whether a notification triggers reconcile, (3) reconcile_init_state, the state that the reconcile_core starts with, (4) reconcile_done, a predicate on whether the reconcile is done.

The simple controller's reconcile_core is currently very simple: At init_pc step, it sends a request to get the CR and moves to after_get_cr_pc; At after_get_cr_pc step, it sends a request to create a configmap subrsource and moves to after_create_cm_pc; At other pc, it simply returns and says the reconcile is done.

Another notable difference from previous examples is that each message now has three fields: src, dst and content. The distributed state machine and the host state machine check if the message is sent to the correct dst host, and the content is actually the old message type -- API request, response, and notification event.

### What does the proof do

The liveness proof in examples/kubernetes_cluster/proof/reconcile_liveness.rs proves "for all CR object, CR exists leads to the configmap subresource exists". Stability is left for the next iteration.

The general proof strategy is as follows:
1. CR object exists ~> CR added event sent
2. CR added event sent ~> simple controller is in reconcile
3. simple controller is in reconcile ~> configmap exists

Prove 2 requires us to consider two cases: (1) simple controller is already in reconcile when CR added event is received (trivial) and (2) simple controller is not in reconcile when CR added event is received (apply WF1 here).

Prove 3 requires us to consider all the intermediate states of reconcile and prove each of them leads to configmap exists.
I use a new proof-rule lemma `leads_to_confluence` here.

More detailed proof steps and explanations can be found in the code.

### Compromises in the spec and proof

I made a few compromises in the spec to be able to deliver a demo proof, including:
1. reconcile_core should check the response and decides the next step. For example, if the response of the get CR request is OK, it should move to the next step to create configmap, otherwise, it should return error.
2. We only talk about creation and deletion request but not update request in the spec because each resource object has an empty body. Solving this might takes a lot of engineering effort and is a must to do for future examples, but not very urgent for this example.
6. The controller spec always reschedule the reconcile when reconcile is done. This makes liveness proof easier but is not very flexible. Well, actually always rescheduling the reconcile is a very good practice to avoid liveness bugs, as long as the time interval is not too short.
7. reconcile_state should be more generic.
8. Kubenetes API spec probably should have multiple different actions for etcd, built-in controllers and API server. Of course, this can make the proof much more complex and it is not a super urgent issue.
9. Stability should be proved. This should not be very difficult.
10. Some of the lemmas cannot be proved without Verus support on FnSpec call trigger. Listed below.

### Unverified proofs
- tla_exists_and/or_equality and tla_forall_and/or_equality (they are not used in the proof yet)
- lemma_always_res_always_exists_implies_all_delete_never_sent
- lemma_exists_get_resp_msg_sent_and_get_cr_leads_to_create_cm

In general, I was struggling with tla_forall/tla_exists involved formulas because I find it difficult to prove this pattern
```
tla_exists(|m| lift_state(|s| ...)) === lift_state(|s| exists |m| ...)
```


Signed-off-by: Xudong Sun <xudongs1@vmware.com>