Revise network model: at-most-once delivery and multiset for in-flight messages (#67)
This PR revises the network model to allow at-most-once delivery for each message.

To make sure messages with the same "payload" does not get dropped by the network model unintentionally, we use a multiset to store all the messages in flight. If a message is sent to the network, the multiset increases the count for this message (the count represents how many messages with the same payload are in flight). If a message is delivered to a host, the counter decrements.

Revising the network model breaks some lemmas because we previously assumed "once a message is in flight, it is always in flight and can be delivered many times" in the proof. This PR also fixes these broken lemmas.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>