Implement simple controller based on the example from kube-rs (#62)
This PR implements an untrusted simple controller that creates a configmap for the custom resource (which matches the spec). It is written in the "state machine" style -- the logic is written in a `reconcile_core` function which is called repeatedly in `reconcile` function. The `reconcile_core` takes the key to the custom resource object, the current local state, the response of the last request, and returns a new local state and a pending request.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>