Specify the controller's list-then-watch pattern (#71)
Specify the controller's list-then-watch pattern.

Controllers use the list-then-watch pattern to observe the cluster state: it first issues a list to read the current cluster state, and the list returns a resource version indicating the "version" of the cluster state the controller just reads. Every single update to any object in the cluster state will increment the resource version number. After the list returns, the controller issues a watch request to continuously receive updates to the cluster state. When issuing the watch request, the controller specifies the resource version number returned by the list, which asks the apiserver to stream updates starting from the resource version.

The current specification is much more simplified than the above: since we have not specified resource version, list will not return any resource version. The watch request is also not specified yet so after the initial list returns the controller will enter a watching state and starts to receive whatever update notifications that are pending in the network.

This liveness proof of the simple controller is broken since the controller spec changes a lot. I will leave the proof there before the spec gets stable and fix all the broken proofs in one batch later.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>