Add channel manager to allocate channel id for each controller request (#86)
Use ChannelManager to wrap around the channel id. ChannelManager is part of the host state machine and allocates a new channel id when sending a new request.

Ideally, the concept of channel should be modeled in the network state machine, but I found that will complicate the spec a lot. For example, we will need to differentiate whether the host is sending to an existing channel or sending to a new channel, and when a new channel is created in the network, the host somehow needs to know the channel id and waits for the message sent back to that channel. We will need to revisit the design of network state machine later.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>