Specify controller crash behavior (#91)
Specify controller crash and add weak fairness assumption to express "the controller eventually stops crashing".

To support specifying crash, we place controller_state in an Option and use Option::None to represent the case where the controller is not running, and we add a crash_enabled bit as the precondition to crash the controller. We add three new actions to the top-level state machine:
- crash_controller that sets the controller state to Option::None and if crash_enabled
- restart_controller that restarts the controller with the initial controller state if the controller is not running
- disable_crash that sets crash_enabled to false

We also add weak fairness assumption on restart_controller and disable_crash so that later we can prove eventually the controller stops crashing and is running alive.

Besides, we fix a few lemmas in cluster.rs. There are more broken lemmas to fix. Most of them need to be proved in a context where the controller is always alive and never crashes, which is not in the scope of this PR.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>