Migrate to Verus main_new
The only changes to *.rs files are to switch imports from pervasive::*
to vstd.

There was an `impl` for `Option` in `api_method.rs` which only worked
because Verus used to put the `pervasive` module inside the crate being
built; now that we use a separate vstd crate, this no longer works.

Signed-off-by: Tej Chajed <tchajed@vmware.com>