Simplify the spec and shim layer implementation using DynamicObject (#99)
Implement DynamicObject and its view to simplify the spec and shim layer.

For the spec, we no longer need to use the KubernetesObject enum (which carries all the pre-defined resource types) in the request/response APIs. Instead, the controller and the Kubernetes API will communicate all kinds of objects in the form of DynamicObject through the network.

For the shim layer, we no longer need to pattern match the request for each Kubernetes resource type because reconcile_core and the shim layer will communicate all kinds of objects in the form of DynamicObject. The shim layer only needs to create an API handle for DynamicObject, but reconcile_core needs to do some extra serialization and deserialization work to translate between DynamicObject and other resource types.

Signed-off-by: Xudong Sun <xudongs1@vmware.com>