Invoke `reconcile_core` with the custom resource object, rather than its reference (#123)
Changes the signature of `reconcile_core` from `reconcile_core(&KubeObjectRef, ...) -> ...` to `reconcile_core(&CustomResource, ...) -> ...` and updates the related spec and proof code accordingly.

The purpose is to simplify the controller implementation. More concretely, the first thing that almost every controller does is to get the triggering custom resource (cr) object. Instead of letting each controller send a request to get the custom resource object and store it in its local reconcile state, our shim layer can directly get the object (by a quorum read at etcd) and pass it to `reconcile_core`.

### Changes in the spec state machine and impacts on the proof

To make the spec state machine match the change in the exec code, this PR also revises the `schedule_controller_reconcile` action to make it pass the cr object to the controller when scheduling reconcile. The new `schedule_controller_reconcile` action matches (and trusts) the behavior that:
(1) `reconcile_with` (in the shim layer) will be triggered when a cr object gets created,
(2) inside `reconcile_with`, the shim layer only calls `reconcile_core` if the cr object exists,
(3) the shim layer passes the cr object to `reconcile_core`, and
(4) the `reconcile_with` gets re-queued unless the cr object no longer exists.

This change also simplifies the liveness proof. Previously one needed to reason about the process of "under the assumption that the cr object always exists, the controller sending a get cr request leads to the controller receiving an OK response containing exactly the same cr object" when proving liveness, but it is no longer needed because now `reconcile_core` gets the cr object "for free".

### What if you get a stale object?

One might wonder how we avoid passing a stale cr object to `reconcile_core`. It depends:
(1) The shim layer does NOT get the cr object from the local cache of API server or the controller, but from a quorum read to etcd (of course we need to trust etcd and Kubernetes here). So we don't worry about stale cache problems.
(2) The "time-of-check to time-of-use" (TOCTOU) issue still exists, and there is no way to prevent that unless we add locks on the cr objects. However, our goal is never to prevent TOCTOU issue but to ensure the controller can survive it and eventually reaches the desired state as long as the cr object gets stable. So what we really care about is whether the spec state machine can capture all the possible TOCTOU issues that could happen in reality. Unfortunately, we don't have a rigorous proof of this yet. An informal way to reason about this is that we have a client host state machine that can create or delete any cr object at any point (later we will also make it able to update the cr object), so for any real-world executions where the cr object gets changed after the shim layer reads it and passes it to `reconcile_core`, our state machine has a corresponding execution that right after `schedule_controller_reconcile` schedules a reconcile with the cr object, the client host changes the cr object.

### New problems introduced

This is actually not a _new_ problem. As we have more and more generic types to carry in our proof, the proof functions read quite verbose. Associated types, once supported, can help solve the problem.

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>