Specify Kubernetes API's update operations (#131)
Revise the Kubernetes API state machine to support handling update requests.

In short, each update request replaces one existing object stored in etcd. The key of handling update in Kubernetes is to check and update the resource version of each object correctly. The resource version is used for multi-version concurrency control in Kubernetes and etcd. Each object has a resource version number and each update to the object increases the resource version. Only the update request that carries the current resource version of the object can successfully update the object.

This versioning mechanism is very important to prove the object created/updated by the controller remains unchanged: suppose the controller always gets the object and then updates the object with the resource version from the get response, the updated object will have a newer (larger) resource version, and all the pending update requests from the previously crashed controller will fail to update this object because their resource version is lower.

An update operation might result in an error response such as (1) Bad request if the namespace or name on the request does not match the object, (2) ObjectNotFound if the object does not exist in etcd, (3) Conflict if the resource version on the request does not match the current resource version of the stored object.

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>