Make from_dynamic_object and unmarshal return Result (#132)
In case that they may receive an object or a value that is an unexpected type or misses certain fields or has fields with unexpected types, we make the two methods `from_dynamic_object()` and `marshal()` return `Result<Self, Error>`.

In exec code, just check whether the `try_parse` succeed. In spec code, we have to reason about different cases. Because I divide the error into two types, in addition to the way of converting k8s objects to dynamic objects, the `Error` returned in all the `else` branches (in the spec code `from_dynamic_object()`) present an alternative sequence of the two types of errors, which is rather tedious.

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>