Prove pending requests with old chan_id eventually are all gone (#138)
This PR proves that from any arbitrary state (`true_pred()`), all the APIRequest messages sent to the Kubernetes API with a chan_id smaller than a specific nat will eventually leave the network.

This lemma is particularly important when proving the object remains in a particular state stably given that update operations are supported in the Kubernetes API state machine. Although it is obvious that the controller always tries to maintain the object in a stable state (by updating the object with the same spec, assuming the CR remains unchanged),
we also need to somehow show that there are no other update requests in the network that can drive the object to a different state -- such requests might come from the previously crashed controller (they crash right after sending out an update request). An important step in the liveness proof here is to prove such requests will eventually leave the network (i.e., all get handled by the Kubernetes API) so they no longer update the object managed by the controller. And after that point the controller will update the object to the desired state and no one will interfere with the object again.

Note that this lemma will become even more useful if later we add the specification for other controllers that might interfere with our verified controller (by updating the object managed by the controller) but eventually stop interfering.

To prove this lemma, we also need some more support from vstd::multiset: https://github.com/verus-lang/verus/pull/579.

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>