Support update operations for rabbitmq controller (#144)
- In the official rabbitmq controller, only scale-up is supported. When
user wants to scale down, rabbitmq controller will reject the request
and return `nil` in reconciliation loop. Currently, in our controller,
an error will be returned.

- For scale up, the official rabbitmq controller will execute a command
`rabbitmq-queues rebalance all`. This command is trying to split
different queues' leader to different nodes, which will utilize
resources more efficiently. However, this is not necessary if we only
want the cluster to perform right functionality.

---------

Signed-off-by: ZichengMa <zma17@illinois.edu>