Prove liveness for the zookeeper controller with update operations (#146)
Proves the liveness property of the zookeeper controller that if the
zookeeper cr always exists and its spec remains unchanged, then the
zookeeper statefulset eventually always exists and its spec remains
unchanged (desired state is matched by the current state).

The key challenge is to prove the statefulset's spec remains unchanged
even with update requests pending in the network that come from the
previously crashed controllers working on a different zookeeper cr's
spec. Our strategy is to first prove that the desired state is achieved
and stable in a context where there are no more disruptive pending
requests, and then prove from `init /\ next /\ wf` the top level cluster
eventually reaches a state that all the disruptive pending requests are
gone.

There are still a few lemmas not proved yet. We will finish proving all
such lemmas in the following PRs.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>