Prove termination of zookeeper controller (#158)
Prove termination of zookeeper controller that when controller will no
longer be restarted (always crash is disabled), the controller will
always eventually enter reconcile idle state. That is basically `spec |=
true ~> reconcile_idle`.

The strategy is to prove for reach case of the controller steps, `spec
|= step ~> reconcile_idle` holds and then combine them together. It's
obvious that there is no pending request at initial step and for other
step (except done and error) there is either pending request or
corresponding response in flight. Then use `wf1` of kubernetes api to
show that in flight request will also lead to a corresponding response
and use `wf1` of controller to show that with a response the controller
with enter the next state.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>