Support unverified third party APIs in controllers (#180)
This PR allows developers to call any (unverified) third-party APIs
besides the Kubernetes client APIs when implementing the controller.

It adds `Lib` as a generic type to `Reconciler`. `Lib` also takes two
generic types `LibInputType` and `LibOutputType`. `Lib` wraps all
interactions with the third-party APIs, and it implements the trait
`ExternalLibrary<LibInputType, LibOutputType>`. `ExternalLibrary` has a
method `process` which calls the third-party API in a synchronous way:
it takes the input from `reconcile_core` (i.e., `LibInputType`) and
returns an output (`Option<LibOutputType>`).

During every reconcile loop, the reconciler will return an
`Option<Request>` (instead of the original `Option<KubeAPIRequest>`) to
the `shim_layer`. `Request` is an enum with two variants for the two
types of effects issued by the controller: invoking Kubernetes client
APIs or third-party APIs. The shim layer will do pattern matching to the
`Request` accordingly, which will either invoke the Kubernetes client
API as before, or invoke the third-party APIs by calling the
developer-provided `process`.

Correspondingly, the input to the reconciler, `Response`, is also an
enum that encodes the two types of responses from the Kubernetes client
APIs or third-party APIs.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>