Support external APIs in the state machine (#185)
The core idea is that we add execution of external functions to the
`continue_reconcile` action.

To support this, we add `lib_response` to the controller state struct.
Then there are three types of actions in `continue_reconcile` which are
categorized based on the return request of `reconcile_core`:
1. If the `reconcile_core` returns a k8s api request, then we set the
`pending_req_msg` field as before;
2. If the `reconcile_core` returns a external request, then we directly
call the external library to process the request, which means we make
the two steps done in one action. Then we set the `lib_response` field
to the returned response.
3. If the `reconcile_core` returns no request, just leave both of the
fields `Option::None`.

We should be careful about the preconditions of `continue_reconcile`. If
the reconcile state has a pending request message, then the
`lib_response` field must be empty because this means for this round of
reconcile, we only process the k8s api response.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>