Implement the reference fluent controller (#187)
This is a simplified version of the fluent-operator:
https://github.com/fluent/fluent-operator. The major purpose of this PR
is to show the feasibility of building a fluent controller using
kube-rs.

In general, this controller (1) creates cluster role, service account,
and cluster role binding to grant access to reading pods in Kubernetes
to the fluent-bit processes (also running as pods), (2) creates secret
to store the configuration of fluent-bit and (3) creates daemon set to
deploy fluent-bit process on each Kubernetes node in the cluster.

Currently, the reference controller does not support update and clean-up
(deletion) operations, and the only field `config` is not actually used.
Future PRs will add more features to this controller.

One major difference from the official fluent operator is about single
v.s. multi reconcile loops. The official fluent operator uses multiple
reconcile loops to manage fluent-bit but our current implementation
manages everything in one reconcile loop. Implementing support for
multiple reconcile loops in Anvil is an important future task.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>