Add simplified specs to the zookeeper external library (#190)
The spec models the data stored in zookeeper as a map whose key is the
zk node path. Calling the zk API to write to the zk node will insert the
key-value pair into the map. The spec is highly simplified compared to
the real zookeeper and is tailored for the use of the zookeeper
controller.

There are still many important problems left for future PRs to deal
with, such as: (1) what is our assumption on the external APIs? Do we
assume they always return the result successfully, or they might return
some transient error (and eventually get back to normal)? Or some other
assumptions? (2) The external API's state and result might depend on the
state of the k8s objects. Like, if the zk pods are not ready then the zk
API won't work. Should we try to specify such dependencies?

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>