Move the external api out of the controller state machine (#192)
Add a field `external_api_state` to the state of the whole system for
describing: (1) what inputs and outputs related to the external api are
not handled (the `in_flight` field); (2) the state of the external api
(`external_api_state` field). It is of type `ExternalAPIState` which
basically wraps a set and the user-defined external api state.

`ExternalComm` is used to represent communication between the external
api and controller. It has two variants: `Input` is computed by
controller and sent to and consumed by external api; `Output` is
generated by external api after handling the input. Each invariant also
carries a nat which represents the id when the input to the external api
is generated.

I don't make it a state machine because it should contain more than one
actions and which action to take can't be determined independently
(simply by the input). So it is neither like the network state machine
which only contains one action nor the other state machines which can
determine the action by the input. But I am considering make the
external api a state machine, whose state is `ExternalAPIType::State`
which is provided by the developer/user, not `ExternalAPIState` which is
provided by me.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>