Allocate uid to each created object in k8s API state machine (#198)
In Kubernetes, each object has an uid (unique id) which is used to
differentiate it from any object that shares the same name and was
previously deleted. For ease of specification and proof, we use nat,
instead of a string (used in real implementation), to represent uid for
each object.

Besides, we also simplify the state machine to make it easier to extend
it with built-in controllers. The transition_by_etcd function now
returns the new k8s api state and the output message. Built-in
controllers later will be specified as other actions of the k8s api
state machine (just like handle_request).

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>