Complete and restructure the liveness proof of rabbitmq controller (#214)
This PR proves the liveness property of the rabbitmq controller.
Different from the previous proof, this time both the liveness property
and the state machine specification become more complex: the liveness
property includes the stateful set object managed by the controller, and
the state machine specification has been enriched a lot with validation
rules, garbage collector, finalizer, and many other features. Besides,
the implementation of the rabbitmq controller has changed: it waits for
the garbage collector to recycle the stateful set of the old custom
resource to avoid accidental downscale (see #211 ). All these factors
make it even more challenging to prove liveness.

### Key challenges

One of the key challenges during the proof is to identify and manage a
large number of invariants. To give some concrete examples, here are two
representative invariants:
1. If stateful set exists, then its owner reference must point to and
only point to the current custom resource.
2. No delete request with the stateful set key will be sent to
Kubernetes API.

The first is used to show that finally stateful set created from old
custom resources (if any) will be deleted, and this is done by the
garbage collector. The second is to verify that if we get the desired
stateful set, it will always exist. The proof of 2 requires 1 because we
need it to show the garbage collector never tries to delete the stateful
set (the gc actions is never enabled).

Notably, though we call them "invariants", neither 1 nor 2 holds from
the beginning of the execution (`init()`). That said, they are not the
safety invariants that hold throughout the entire execution, but they
start to hold at some point during the execution and continue to hold
since then.

Proving the liveness property requires a large number of such
invariants, so how to prove and use them becomes a significant part of
the proof burden. Note that the same problem exists even before this PR.

### Proof structure

We notice that a common pattern in the liveness proof is that, to make
the desired property eventually happens, the environment needs to first
enter a state where there are no more perturbations that can interrupt
the execution moving toward the desired property. For example, if we
want the controller to make progress to the desired state, then we need
to first prove that the controller stops crashing from a point.

That said, the entire liveness proof, similar to the actual execution,
is divided into phases. Starting from the initial phase (the initial
state of the state machine), the system first enters a phase where
perturbations (e.g., staleness, crashes, disturbing requests) no longer
exist, then moves to the final phase where the desired state is achieved
by the controller's correct execution. At each phase, we derive some
invariants (the properties that hold from this point) which will be used
to prove the properties of the next phase. Naturally, the invariants
used in the proof can be grouped by the phase. They were referred to as
`assumptions`, `invariants_since_rest_id`,
`invariants_led_to_by_rest_id` (these variable names are indeed
confusing, and we stop calling them in this way).

In this proof, we divide the execution into different phase and
invariants into several groups $I_1,I_2,\cdots,I_n$. More details in
[this
comment](https://github.com/vmware-research/verifiable-controllers/pull/214#issuecomment-1681302812).

In this structure, we often need to prove lemmas in the form of `spec |=
true ~> []p`. If the property `p` talks about messages sent after a
particular point, we often have to choose a `rest_id` and prove that
eventually request messages before `req_id` are gone. This PR introduces
a
[lemma](https://github.com/vmware-research/verifiable-controllers/blob/f5783e1a014e7eedc76e38bfd82fa61de6d4cb98/src/kubernetes_cluster/proof/kubernetes_api_liveness.rs#L276C14-L276C75)
in the `kubernetes_api_liveness` module to help reason about this kind
of invariants, and the controller developers no longer need to directly
reason about different `rest_id`s (And it's indeed better to avoid it).
This improves both the reusability and the readability of the
proof/lemmas.

This PR does not try to finish proving all the invariants but focuses on
ensuring that:
1. The invariants are grouped correctly into phases that follow the same
logical order of the proof/execution.
2. The invariants should be provable with all antecedent invariants
required.

The phases of the invariants used in the liveness proof are:

- 0: Invariants that hold throughout the entire execution (i.e., safety
invariants).
- 1: Crash and K8s busy are disabled, and the object that is in the
schedule for reconciliation is updated.
- 2: The object being reconciled is updated.
- 3: Every creation and update request to the same object has the same
effect.
- 4: Objects only have their owner references pointing to the current
cr, not the previously-deleted cr.
- 5: No deletion requests are sent to the K8s API.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>